[camera]merge all properties into a single file
