--- packages/camera/camera/ios/Classes/VideoFormat.h
@@ -1,14 +1,0 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-#import <Foundation/Foundation.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-/**
- * Gets VideoFormat from its string representation.
- */
-extern OSType FLTGetVideoFormatFromString(NSString *videoFormatString);
-
-NS_ASSUME_NONNULL_END
